+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Added marker file for PEP 561. This will allow type checking tools in dependent projects
   to use type annotations from Python-RSA
   ([#136](https://github.com/sybrenstuvel/python-rsa/pull/136)).
+- Use the Chinese Remainder Theorem when decrypting with a private key. This
+  makes decryption 2-4x faster
+  ([#163](https://github.com/sybrenstuvel/python-rsa/pull/163)).
 
 ## Version 4.7.2 - released 2021-02-24